Add a data module
This is a measure for computer information, it goes down to bits but
the internal value is the octet.

We decided against using the byte as the internal value because its
size is architecture-dependent, on DSPs for example a byte could be
16 bits instead of 8 as it is usually understood.  The octet on the
other hand is always defined as eight bits.